changelog refactoring MoviesRemoteMediator.kt and remove same methods to another classes add two classes MoviesNetwork.kt and MoviesRepoDatabase.kt with methods from MoviesRemoteMediator.kt add animate between list of movies and details movie
