+++ app/src/main/res/layout/fragment_movie_details.xml
@@ -1,17 +1,8 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-<com.google.android.material.card.MaterialCardView
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:cardCornerRadius="8dp"
-    app:cardElevation="6dp"
-    app:cardUseCompatPadding="true"    >
-
--->
+<?xml version="1.0" encoding="utf-8"?>
 
 <ScrollView xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:transitionName="trans1"
     android:id="@+id/scroll_details"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
@@ -24,7 +15,6 @@
         android:background="@color/sapphire_blue_one">
 
         <ProgressBar
-            android:transitionName="trans2"
             android:id="@+id/progressBar"
             style="?android:attr/progressBarStyleLarge"
             android:layout_width="wrap_content"
@@ -36,7 +26,6 @@
             app:layout_constraintBottom_toBottomOf="parent" />
 
         <androidx.recyclerview.widget.RecyclerView
-            android:transitionName="trans3"
             android:id="@+id/list_actors"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -48,7 +37,6 @@
             tools:listitem="@layout/actor_item_view_holder" />
 
             <TextView
-                android:transitionName="trans4"
                 android:id="@+id/head_cast_text_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -63,7 +51,6 @@
                 tools:text="@string/head_cast" />
 
             <TextView
-                android:transitionName="trans5"
                 android:id="@+id/all_actors_text_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -79,7 +66,6 @@
                 tools:text="@string/all_actors" />
 
             <TextView
-                android:transitionName="trans6"
                 android:id="@+id/description_text_view"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -98,7 +84,6 @@
                 tools:text="@string/text_description" />
 
             <TextView
-                android:transitionName="trans7"
                 android:id="@+id/head_description_text_view"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -115,7 +100,6 @@
                 tools:text="@string/head_description" />
 
             <TextView
-                android:transitionName="trans8"
                 android:id="@+id/reviews_text_view"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -128,12 +112,8 @@
                 app:layout_constraintStart_toEndOf="@id/reviews_rating_bar"
                 app:layout_constraintTop_toTopOf="@id/reviews_rating_bar"
                 tools:text="125 REVIEWS" />
-        <!--
-                    android:text="@{@string/reviews_text(movie.reviews)}"
-        -->
 
         <RatingBar
-            android:transitionName="trans9"
             android:id="@+id/reviews_rating_bar"
             style="?android:attr/ratingBarStyleSmall"
             android:layout_width="wrap_content"
@@ -149,7 +129,6 @@
                 tools:rating="4" />
 
             <TextView
-                android:transitionName="trans10"
                 android:id="@+id/tag_text_view"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -165,7 +144,6 @@
                 tools:text="@string/tag_text" />
 
             <TextView
-                android:transitionName="@string/fragmentB"
                 android:id="@+id/name_movie_text_view"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
@@ -186,7 +164,6 @@
                 tools:text="@string/name_movie" />
 
             <View
-                android:transitionName="trans12"
                 android:id="@+id/mask_view"
                 android:layout_width="0dp"
                 android:layout_height="0dp"
@@ -199,7 +176,6 @@
                 app:layout_constraintTop_toTopOf="parent" />
 
             <ImageView
-                android:transitionName="trans13"
                 android:id="@+id/movie_image_view"
                 android:layout_width="0dp"
                 android:layout_height="0dp"
@@ -217,7 +193,6 @@
                 tools:src="@drawable/ic_no_image" />
 
             <View
-                android:transitionName="trans14"
                 android:id="@+id/pg_view"
                 android:layout_width="24dp"
                 android:layout_height="24dp"
@@ -231,7 +206,6 @@
                 app:layout_constraintTop_toBottomOf="@id/back_text_view" />
 
             <TextView
-                android:transitionName="trans15"
                 android:id="@+id/pg_text_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -245,7 +219,6 @@
                 tools:text="@string/pg_text" />
 
             <ImageView
-                android:transitionName="trans16"
                 android:id="@+id/back_image_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -260,7 +233,6 @@
                 tools:src="@drawable/ic_navigate_before_12" />
 
             <TextView
-                android:transitionName="trans17"
                 android:id="@+id/back_text_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -279,5 +251,4 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-</ScrollView><!--
-</com.google.android.material.card.MaterialCardView>-->
+</ScrollView>
