+++ app/src/main/res/layout/movie_item_view_holder.xml
@@ -10,7 +10,6 @@
     android:background="@drawable/bg_corners_gradient_stroke">
 
     <ProgressBar
-        android:transitionName="trans1"
         android:id="@+id/progressBar"
         style="?android:attr/progressBarStyle"
         android:layout_width="wrap_content"
@@ -22,7 +21,6 @@
         app:layout_constraintBottom_toBottomOf="parent" />
 
     <TextView
-        android:transitionName="trans2"
         android:id="@+id/minute_text_view"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -41,7 +39,6 @@
         tools:text="137 min" />
 
     <TextView
-        android:transitionName="fragment_a"
         android:id="@+id/name_movie_text_view"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -58,7 +55,6 @@
         tools:text="Avengers: End Game" />
 
     <com.google.android.material.imageview.ShapeableImageView
-        android:transitionName="trans4"
         android:id="@+id/movie_image_view"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -77,7 +73,6 @@
 
 
     <ImageView
-        android:transitionName="trans5"
         android:id="@+id/mask_view"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -90,7 +85,6 @@
 
 
     <TextView
-        android:transitionName="trans6"
         android:id="@+id/reviews_text_view"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -107,7 +101,6 @@
         tools:text="125 REVIEWS" />
 
     <RatingBar
-        android:transitionName="trans7"
         android:id="@+id/reviews_rating_bar"
         style="?android:attr/ratingBarStyleSmall"
         android:layout_width="wrap_content"
@@ -127,7 +120,6 @@
         tools:rating="4" />
 
     <TextView
-        android:transitionName="trans8"
         android:id="@+id/tag_text_view"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -145,7 +137,6 @@
         tools:text="@string/tag_text" />
 
     <View
-        android:transitionName="trans9"
         android:id="@+id/pg_view"
         android:layout_width="24dp"
         android:layout_height="24dp"
@@ -157,7 +148,6 @@
         app:layout_constraintTop_toTopOf="@id/movie_image_view" />
 
     <TextView
-        android:transitionName="trans10"
         android:id="@+id/pg_text_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -173,7 +163,6 @@
         tools:text="@string/pg_text" />
 
     <ImageView
-        android:transitionName="trans11"
         android:id="@+id/favorite"
         android:layout_width="16dp"
         android:layout_height="14dp"
